First audio generation with CIS data simlation

+++ CM7/Middlewares/Inc/synth.h
@@ -45,7 +45,7 @@ extern volatile uint32_t synth_process_cnt;
 /* Exported functions prototypes ---------------------------------------------*/
 int32_t synth_IfftInit(void);
 int32_t synth_PlayInit(void);
-int32_t synth_GetAudioData(uint32_t index);
+int16_t synth_GetAudioData(uint32_t index);
 int32_t synth_GetImageData(uint32_t index);
 int32_t synth_SetImageData(uint32_t index, uint16_t value);
 void synth_AudioProcess(synthModeTypeDef mode);
